requirements.txt
streamlit
opencv-python
numpy
pillow
PyMuPDF
pdf2image
pytesseract
spacy==3.7.2
google-cloud-vision
